# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MNT Embark API — luxury tour company
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /activities/{slug}:
    get:
      operationId: getActivityBySlug
      tags: [activities]
      summary: Get a single activity by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Activity detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityDetail"
        "404":
          description: Activity not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Activity administration ───────────────────────────────────────────
  # Separate from GET /activities, which returns filter groups shaped for the
  # public sidebar: only filterable activities, with live tour counts and no
  # editing fields. These return the rows as stored.
  /admin/activity-groups:
    get:
      operationId: listActivityGroups
      tags: [activities]
      summary: List every activity group, including empty ones
      responses:
        "200":
          description: Activity groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminActivityGroup"
    post:
      operationId: createActivityGroup
      tags: [activities]
      summary: Create an activity group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityGroupInput"
      responses:
        "201":
          description: Created group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminActivityGroup"
        "409":
          description: Slug already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/activity-groups/{id}:
    patch:
      operationId: updateActivityGroup
      tags: [activities]
      summary: Update an activity group
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityGroupUpdate"
      responses:
        "200":
          description: Updated group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminActivityGroup"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Slug already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteActivityGroup
      tags: [activities]
      summary: Delete an activity group
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Group still has activities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/activities:
    get:
      operationId: listAllActivities
      tags: [activities]
      summary: List every activity as stored, filterable or not
      responses:
        "200":
          description: Activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminActivity"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Create an activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityInput"
      responses:
        "201":
          description: Created activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminActivity"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Slug already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/activities/{id}:
    patch:
      operationId: updateActivity
      tags: [activities]
      summary: Update an activity
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityUpdate"
      responses:
        "200":
          description: Updated activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminActivity"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Slug already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteActivity
      tags: [activities]
      summary: Delete an activity
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Activity is still attached to tours
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Tours ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Itinerary Step ─────────────────────────────────────────────────
...
    AdminActivityGroup:
      type: object
      properties:
        id:
          type: number
        slug:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        coverImage:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        selectionMode:
          type: string
          enum: [single, multiple]
        displayOrder:
          type: number
        activityCount:
          type: number
      required:
        [
          id,
          slug,
          name,
          description,
          coverImage,
          icon,
          selectionMode,
          displayOrder,
          activityCount,
        ]
    ActivityGroupInput:
      type: object
      properties:
        slug:
          type: string
          minLength: 2
        name:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        coverImage:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        selectionMode:
          type: string
          enum: [single, multiple]
        displayOrder:
          type: number
      required: [slug, name]
    ActivityGroupUpdate:
      type: object
      properties:
        slug:
          type: string
          minLength: 2
        name:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        coverImage:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        selectionMode:
          type: string
          enum: [single, multiple]
        displayOrder:
          type: number
    AdminActivity:
      type: object
      properties:
        id:
          type: number
        groupId:
          type: number
        groupName:
          type: string
        slug:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        coverImage:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        aliases:
          type: array
          items:
            type: string
        isFilterable:
          type: boolean
        isIndexable:
          type: boolean
        displayOrder:
          type: number
        attractionCount:
          type: number
      required:
        [
          id,
          groupId,
          groupName,
          slug,
          name,
          description,
          coverImage,
          icon,
          aliases,
          isFilterable,
          isIndexable,
          displayOrder,
          attractionCount,
        ]
    ActivityInput:
      type: object
      properties:
        groupId:
          type: number
        slug:
          type: string
          minLength: 2
        name:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        coverImage:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        aliases:
          type: array
          items:
            type: string
        isFilterable:
          type: boolean
        isIndexable:
          type: boolean
        displayOrder:
          type: number
      required: [groupId, slug, name]
    ActivityUpdate:
      type: object
      properties:
        groupId:
          type: number
        slug:
          type: string
          minLength: 2
        name:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        coverImage:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        aliases:
          type: array
          items:
            type: string
        isFilterable:
          type: boolean
        isIndexable:
          type: boolean
        displayOrder:
          type: number

    # ── Tour guide copy ───────────────────────────────────────────────
...
    ActivityDetail:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        coverImage:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        groupSlug:
          type: string
        groupName:
          type: string
        isIndexable:
          type: boolean
        redirectToSlug:
          type: ["string", "null"]
      required: [slug, name, description, coverImage, icon, groupSlug, groupName, isIndexable, redirectToSlug]
...